+++ common/wallets/protocol-format-mapping.js
@@ -6,6 +6,12 @@
  */
 
 /**
+ * List all credential formats supported by the platform.
+ * Used for format-agnostic protocols (e.g. Interact URL).
+ */
+export const ALL_CREDENTIAL_FORMATS = ['ldp_vc', 'jwt_vc_json', 'mso_mdoc'];
+
+/**
  * Default mapping of protocols to supported credential formats.
  * Wallets can override/extend these defaults in their configurations.
  */
@@ -18,10 +24,10 @@ export const PROTOCOL_FORMAT_MAPPING = {
   '18013-7-Annex-C': ['mso_mdoc'],
   '18013-7-Annex-D': ['mso_mdoc'],
 
-  // Todo: These methods could also support the other formats,
+  // Todo: vcapi and chapi could also support the other formats,
   // but unclear exactly how yet
   vcapi: ['ldp_vc'],
-  interact: ['ldp_vc'],
+  interact: ALL_CREDENTIAL_FORMATS,
   chapi: ['ldp_vc'] // 'chapi' is shorthand for the 'web credential request'
   // request format including a VerifiablePresentationRequest query with
   // QueryByExample. This is the predominant method of passing a query through