+++ common/wallets/protocols/interaction.js
@@ -5,22 +5,24 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
+import {ALL_CREDENTIAL_FORMATS} from '../protocol-format-mapping.js';
+
 export const interactionWallet = {
   id: 'interaction',
   name: 'Interaction',
   nameKey: 'protocols_interact_name',
   descriptionKey: 'protocolWallet_interact_description',
-  supportedFormats: ['ldp_vc'],
+  supportedFormats: ALL_CREDENTIAL_FORMATS,
   supportedProtocols: {
     interact: {
       copy: {
         descriptionKey: 'protocolWallet_interact_copy',
-        formats: ['ldp_vc']
+        formats: ALL_CREDENTIAL_FORMATS
         // getUrl defaults to getUrlDefault in helper function
       },
       qr: {
         descriptionKey: 'protocolWallet_interact_qr',
-        formats: ['ldp_vc']
+        formats: ALL_CREDENTIAL_FORMATS
         // getUrl defaults to getUrlDefault in helper function
       }
     }