+++ test/bedrock/180-exchange-protocols.test.js
@@ -89,6 +89,12 @@ describe('Exchange Protocols', () => {
         'profile%3DOID4VP-draft18');
       expect(exchange.protocols['OID4VP-1.0']).to.contain(
         'profile%3DOID4VP-1.0');
+      expect(exchange.protocols.OID4VP).to.contain(
+        'request_uri_method=post');
+      expect(exchange.protocols['OID4VP-1.0']).to.contain(
+        'request_uri_method=post');
+      expect(exchange.protocols['OID4VP-draft18']).to.not.contain(
+        'request_uri_method');
     });
 
     it('should include protocols object for vc-api workflow', async () => {